create components via cli
