+++ applications/tari_validator_node/tests/features/basic.feature
@@ -24,14 +24,17 @@ Feature: Basic scenarios
     Then the validator node VAL_1 is listed as registered
     # Then the validator node VAL_2 is listed as registered
 
+    # Register the "counter" template
+    When validator node VAL_1 registers the template "counter"
+    When miner MINER mines 20 new blocks
+    Then the template "counter" is listed as registered by the validator node VAL_1
+    # Then the template "counter" is listed as registered by the validator node VAL_2
+
     # Create an account
     When I create an account ACC_1 on VAL_1
 
-    # Register the "counter" template
-    # When validator node VAL_1 registers the template "counter"
-    # When miner MINER mines 20 new blocks
-    # Then the template "counter" is listed as registered by the validator node VAL_1
-    # Then the template "counter" is listed as registered by the validator node VAL_2
+    # Create a new Counter component
+    When I create a component COUNTER_1 of template "counter" on VAL_1 using "new"
 
     # Call the constructor in the "counter" template
     # When the validator node VAL_1 calls the constructor "new" on the template "counter" to create component COUNTER_1
